Fix 400 error & typos
The repo has been updated with Alsid information, plus some typos've
been fixed and, more importantly, the 400 error we had upon linking a
new work item and creating it at the same time has been fixed.

This 400 error was due to the Bug template we're using on azdo's side,
where we redefined the Severity field's values. None matched the strings
in this plugin, hence the error. Modifying the strings in this plugin to
match the strings in our azdo template makes it work©.