Fix fails to call request function
I still don't understand everything, but the "execute.query" and
"execute.response" events are registered in app.js and modal.js, and
both are executed upon a trigger on respectively those.
However, both are sharing the local storage (same
setMessageArg/getMessageArg pair of functions) and clear it just after
getting the value. So the second one to grab it fails, which is what
happens in prod.

Applying and checking for null values respectively after and before use
looks to do the trick, hence this commit.